Fix most warnings
Mostly signed/unsigned comparisons, unused variables, and shadow
variables.

The only files not yet warning-free are qman_driver.c and
bman_driver.c, and this is because fixing them would just complicate
efforts to re-sync those with upstream driver sources (which could
well fix the warnings anyway).

+++ apps/include/bigatomic.h
@@ -50,7 +50,7 @@ static inline u64 bigatomic_read(const struct bigatomic *b)
 	do {
 		upper = atomic_read(&b->upper);
 		lower = atomic_read(&b->lower);
-	} while (upper != atomic_read(&b->upper));
+	} while (upper != (u32)atomic_read(&b->upper));
 	return ((u64)upper << 32) | (u64)lower;
 }
 static inline void bigatomic_inc(struct bigatomic *b)
